pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptc"
version = "0.1.0"
description = "Measure worst-case context exposure in your Claude / Cursor setup."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Eden" }]
keywords = ["claude", "cursor", "llm", "prompt", "context", "tokens"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "tiktoken>=0.7",
    "markdown-it-py>=3.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
promptc = "promptc.cli:main"

[project.urls]
Homepage = "https://github.com/Eden/promptc"
Repository = "https://github.com/Eden/promptc"
Issues = "https://github.com/Eden/promptc/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/promptc"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
